% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls.R
\name{bls_earnings}
\alias{bls_earnings}
\title{Average Hourly Earnings (private sector, seasonally adjusted)}
\usage{
bls_earnings(start_year = NULL, end_year = NULL, key = NULL)
}
\arguments{
\item{start_year}{Start year}

\item{end_year}{End year}

\item{key}{Optional API key}
}
\value{
tibble: series_id, year, period, period_name, value, date
}
\description{
Average Hourly Earnings (private sector, seasonally adjusted)
}
