% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls.R
\name{bls_ppi}
\alias{bls_ppi}
\title{Producer Price Index (all commodities)}
\usage{
bls_ppi(start_year = NULL, end_year = NULL, key = NULL)
}
\arguments{
\item{start_year}{Start year}

\item{end_year}{End year}

\item{key}{Optional API key}
}
\value{
tibble: series_id, year, period, period_name, value, date
}
\description{
Producer Price Index (all commodities)
}
